moved getter and setter of traceId and parentId into EventBean, followed w3c trace context specification, removed ramsey/uuid dependency
