+++ composer.json
@@ -5,7 +5,6 @@
     "require" : {
         "php" : ">= 7.1",
         "guzzlehttp/guzzle": "6.*",
-        "ramsey/uuid": "^3.7",
         "ralouphie/getallheaders": "2.0.5",
         "ext-curl" : "*",
         "ext-json": "*"
